==> scripts/01-downloadData.Rmd <==
---
title: "Download UKBB data for the Analysis"
author: "melike"
date: "30 July 2018"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,cache=T, message=F, warning=F, tidy=TRUE, tidy.opts=list(width.cutoff=55))
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

# Create the necesarry folders

```{bash}
mkdir -p data/raw/ukbb
```

# Download ukb10904.enc file

This file contains the encoded data. To download, we need to go to application specific data showcase, downloads section on UK Biobank website and download the file after entering the md5 checksum sent us through e-mail. This file is downloaded and saved under './data/raw/ukbb/ukb10904.enc'

We also need the key file sent through e-mail. That also saved as './data/raw/ukbb/k3068.key'

# Download helper programs

```{bash}
mkdir data/raw/ukbb/helpers
cd data/raw/ukbb/helpers
wget  -nd  "biobank.ndph.ox.ac.uk/showcase/util/ukbmd5"
wget  -nd  "biobank.ndph.ox.ac.uk/showcase/util/ukbunpack"
wget  -nd  "biobank.ndph.ox.ac.uk/showcase/util/ukbconv"
wget  -nd  "biobank.ndph.ox.ac.uk/showcase/util/ukbgene"
wget  -nd  "biobank.ndph.ox.ac.uk/showcase/util/encoding.ukb"
```

# Decrypting the file

## Verify the integrity of file through md5 

This command should give the same md5 checksum we used to download the data.

```{bash}
./data/raw/ukbb/helpers/ukbmd5 ./data/raw/ukbb/ukb10904.enc
```

It is the same, so we can continue.

## Decompress file

```{bash}
cd ./data/raw/ukbb/
mykey=$(head -1  k3068.key )
./helpers/ukbunpack ukb10904.enc $mykey
```

```{bash}
cd ./data/raw/ukbb/
./helpers/ukbconv ukb10904.enc_ukb csv
```

# Download other helper files

```{bash}
mkdir -p ./data/raw/ukbb/helperfiles
cd ./data/raw/ukbb/helperfiles
wget -nd -O field.txt "biobank.ndph.ox.ac.uk/showcase/scdown.cgi?fmt=txt&id=1"
wget -nd -O encoding.txt "biobank.ndph.ox.ac.uk/showcase/scdown.cgi?fmt=txt&id=2"
wget -nd -O instances.txt "biobank.ndph.ox.ac.uk/showcase/scdown.cgi?fmt=txt&id=9"
wget -nd -O archive.txt "biobank.ndph.ox.ac.uk/showcase/scdown.cgi?fmt=txt&id=4"
wget -nd -O category.txt "biobank.ndph.ox.ac.uk/showcase/scdown.cgi?fmt=txt&id=3"
wget -nd -O catbrowse.txt "biobank.ndph.ox.ac.uk/showcase/scdown.cgi?fmt=txt&id=13"
wget -nd -O schema.txt "biobank.ndph.ox.ac.uk/showcase/scdown.cgi?fmt=txt&id=999"
wget -nd -O ehierint.txt "biobank.ndph.ox.ac.uk/showcase/scdown.cgi?fmt=txt&id=11"
wget -nd -O ehierstring.txt "biobank.ndph.ox.ac.uk/showcase/scdown.cgi?fmt=txt&id=12"
wget -nd -O insvalue.txt "biobank.ndph.ox.ac.uk/showcase/scdown.cgi?fmt=txt&id=10"
wget -nd -O esimpdate.txt "biobank.ndph.ox.ac.uk/showcase/scdown.cgi?fmt=txt&id=8"
wget -nd -O esimpint.txt "biobank.ndph.ox.ac.uk/showcase/scdown.cgi?fmt=txt&id=5"
wget -nd -O esimpreal.txt "biobank.ndph.ox.ac.uk/showcase/scdown.cgi?fmt=txt&id=7"
wget -nd -O esimpstring.txt "biobank.ndph.ox.ac.uk/showcase/scdown.cgi?fmt=txt&id=6"
```

# Download data coding files

Some of the fields in data are encoded. The coding files include definitions. These are downloaded from the webpage instead of using wget simply because there is no link to the files but downloading only works by clicking on download button on the webpages. The relevant coding id is retrieved from ./data/raw/ukbb/helperfiles/encoding.txt and files are saved under ./data/raw/ukbb/datacoding


# Download fam files

```{bash}
mkdir -p data/raw/ukbb/fam
cd ./data/raw/ukbb/fam
echo 30688 | cat - ../k3068.key > .ukbkey
for i in {1..22}
do
  ../helpers/ukbgene cal -c$i -m
done
```

# Download sample files

```{bash}
mkdir -p data/raw/ukbb/sample
cd ./data/raw/ukbb/sample
echo 30688 | cat - ../k3068.key > .ukbkey
for i in {1..22}
do
  ../helpers/ukbgene imp -c$i -m
done
```

# Downloaad relatedness

```{bash}
mkdir -p data/raw/ukbb/rel
cd ./data/raw/ukbb/rel
echo 30688 | cat - ../k3068.key > .ukbkey
../helpers/ukbgene rel
```

# Download resources

```{bash}
cd ./data/raw/ukbb/
wget "http://biobank.ndph.ox.ac.uk/showcase/auxdata/ukb_snp_qc.txt"
```
